+++ agents-run-api/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @inkeep/agents-run-api
 
+## 0.33.2
+
+### Patch Changes
+
+- 4b2fd62: tool history perserved
+- Updated dependencies [4b2fd62]
+- Updated dependencies [bbbed5e]
+  - @inkeep/agents-core@0.33.2
+
+## 0.33.1
+
+### Patch Changes
+
+- 98f139a: Updated agent cil
+- Updated dependencies [98f139a]
+  - @inkeep/agents-core@0.33.1
+
 ## 0.33.0
 
 ### Minor Changes